Add animations

+++ Web-Technologies/Full-Stack-Web-Development-with-React-Specialization/3.Multiplatform-Mobile-App-Development-with-React-Native/confusion/components/FavoriteComponent.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
+import * as Animatable from 'react-native-animatable';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
@@ -30,19 +31,40 @@ class Favorites extends Component {
     render() {
 
         const { navigate } = this.props.navigation;
-        const rightButton = [
-            {
-                text: 'Delete', 
-                type: 'delete',
-                onPress: () => this.props.deleteFavorite(item.id)
-            }
-        ];
+       
 
 
         const renderMenuItem = ({item, index}) => {
     
+            const rightButton = [
+                {
+                    text: 'Delete', 
+                    type: 'delete',
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                        
+                    }
+                }
+            ];
+
             return (
                 <Swipeout right={rightButton} autoClose={true}>
+                    <Animatable.View animation="fadeInRightBig" duration={2000}>                
                     <ListItem
                         key={index}
                         title={item.name}
@@ -51,6 +73,7 @@ class Favorites extends Component {
                         onPress={() => navigate('Dishdetail', { dishId: item.id })}
                         leftAvatar={{ source: {uri: baseUrl + item.image}}}
                         />
+                </Animatable.View>
                 </Swipeout>
             );
         };
@@ -69,6 +92,7 @@ class Favorites extends Component {
         }
         else {
             return (
+
                 <FlatList 
                 data={this.props.dishes.dishes.filter(dish => this.props.favorites.some(el => el === dish.id))}
                     renderItem={renderMenuItem}
